Fixed the Kinematic class and updated all classes with a composition relation. Re-factored all other classes, finished Basic_Motion_Controller, and debugged system.
Note: Operates, but needs a little tweeking to fine tune the movement of
the turtle and to fix the end condition in the Level object inside
Basic_Motion_Control.java.

+++ basic_motion/src/ai/Movement.java
@@ -18,6 +18,6 @@ public abstract class Movement {
 	 * @param time the elapsed time since the previous 
 	 * update.
 	 */
-	public abstract void update(Kinematic kin, int time);
+	public abstract void update(Kinematic kin, double time);
 	
 }
